Update DMC URL
