+++ lib/dmc/fsr_configuration.rb
@@ -1,31 +1,11 @@
 # frozen_string_literal: true
 
-module DMC
-  class FSRConfiguration < Common::Client::Configuration::REST
-    def self.base_request_headers
-      super.merge(
-        'client_id' => Settings.dmc.client_id,
-        'client_secret' => Settings.dmc.client_secret
-      )
-    end
+require 'dmc/base_configuration'
 
+module DMC
+  class FSRConfiguration < DMC::BaseConfiguration
     def service_name
       'FSR'
-    end
-
-    def base_path
-      "#{Settings.dmc.url}/api/v1/digital-services/financial-status-report/"
-    end
-
-    def connection
-      Faraday.new(base_path, headers: base_request_headers, request: request_options) do |f|
-        f.use :breakers
-        f.use Faraday::Response::RaiseError
-        f.request :json
-        f.response :betamocks if Settings.dmc.mock_fsr
-        f.response :json
-        f.adapter Faraday.default_adapter
-      end
     end
   end
 end
